+++ Magic/CHANGELOG.md
@@ -5,6 +5,7 @@
  - Broomsticks will upgrade on use in the potter configs, too
  - Heroes integration improvements:
    - Allow using disabled icons and skull-based disabled icons.
+   - Disabled icons will show in skill selector for unavailable skills
 
 ## 6.6